requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
minio==7.2.0
requests==2.31.0
python-dotenv==1.0.0
fastapi
uvicorn[standard]==0.30.1
